+++ tagpack/__init__.py
@@ -1,0 +1,3 @@
+"""Module config for tagpack-tool"""
+
+from tagpack._version import __version__